Remove fileDownload component. Just use the mixin
